refactor: merge 6 tropical test files into single test_tropical.py
